Some experimentation with expression structure.
